🦄 refactor(components): 组件搬迁与重命名
新的目录结构：
src/components/
├── canvas/                    # 画布核心
│   ├── CanvasStage.vue       # 主画布
│   ├── SelectionOverlay.vue  # 选区覆盖层
│   └── layers/               # 图层组件
│       ├── CircleLayer.vue
│       ├── GroupLayer.vue
│       ├── ImageLayer.vue
│       ├── RectLayer.vue
│       └── TextLayer.vue
├── ui/                        # UI 组件
│   ├── panels/               # 固定面板
│   │   ├── AppHeader.vue     # 应用头部
│   │   ├── ToolPanel.vue     # 工具面板
│   │   └── InspectorPanel.vue # 属性检查器
│   ├── floating/             # 浮动控件
│   │   ├── ContextMenu.vue   # 右键菜单
│   │   ├── HoverToolbar.vue  # 悬浮工具栏
│   │   └── GroupActions.vue  # 组合操作
│   └── common/               # 通用组件
│       └── ImageMenu.vue     # 图片菜单
└── composables/              # Composables (已创建)
