+++ src/components/canvas/CanvasStage.vue
@@ -65,8 +65,8 @@ import CircleLayer from './layers/CircleLayer.vue';
 import ImageLayer from './layers/ImageLayer.vue';
 import GroupLayer from './layers/GroupLayer.vue';
 import SelectionOverlay from './SelectionOverlay.vue';
-import ContextToolbar from '../layout/ContextToolbar.vue';
-import GroupToolbar from './GroupToolbar.vue';
+import ContextToolbar from '../ui/floating/HoverToolbar.vue';
+import GroupToolbar from '../ui/floating/GroupActions.vue';
 import { ToolManager } from '@/core/ToolManager';
 import { NodeFactory } from '@/core/services/NodeFactory';
 import { GroupService } from '@/core/services/GroupService';
