+++ src/components/canvas/SelectionOverlay.vue
