+++ src/components/canvas/layers/CircleLayer.vue
