+++ src/components/canvas/layers/GroupLayer.vue
