+++ src/components/canvas/layers/ImageLayer.vue
