+++ src/components/canvas/layers/RectLayer.vue
