+++ src/components/canvas/layers/TextLayer.vue
