+++ src/components/ui/floating/GroupActions.vue
