+++ src/components/ui/panels/AppHeader.vue
